bash script: It now obtains the end position of the structural variant instead of the filter. It also now filters on lengths instead of the python script doing it
Python script:
Added the function that gets the amount of SVs that overlap with a non-syntenic region
It now doesn't check the length.
New method of checking the type of SV.
Changed some variable names.
The barplot now also shows the amount of non-syntenic SVs.
Updated the docstrings and comments.

Updated the markdown file.